requirements.txt
requests
pypdf
pdfplumber
psycopg2-binary
beautifulsoup4